+++ CHANGELOG.markdown
@@ -4,8 +4,7 @@ Google Analytics Server Side Changelog
 Version 0.7.6 Beta
 ------------------
 
-- Throw exceptions when browsercap ini file not loaded but specified
-- Inform user via php error log when browsercap ini file has been updated and needs loading
+- BrowserCap ini file is now parsed and dealt with by the code rather than using php's built in get_browser. 
 - BrowserCap ini file now updated whenever update available on server
 
 Version 0.7.5 Beta
@@ -13,7 +12,7 @@ Version 0.7.5 Beta
 
 - Fix wrong name variable used in BrowserCap
 - Ensure UserAgent has been sent to BotInfo
-- Ensure AcceptedLanguage/RemoteAddress/UserAgent has been sent to Http
+- Ensure AcceptedLanguage / RemoteAddress / UserAgent has been sent to Http
 
 Version 0.7.4 Beta
 ------------------